Detect errors in lzma_code and fail at that point.
We now write full buffers correctly.  When the buffer is full -
i.e. avail_out == 0:
 1: Write the current buffer.
 2: Reset the output buffer pointer and signal available space.
 3: call lzma_code again.
 4: If avail_out == 0 again, loop back to 1

This minimises the number of file writes, and makes the process
read driven: i.e. the main for loop in encode could be extracted
as a single function call with a string buffer to write.  That's
what I want for the xz lib, so that's progress.