Most of the way there with writing a reader-driven decompressor.
The objective is to let an external caller ask for a string similar to
fgets - fill either len characters or to the next newline, whichever is
less.  This then pulls data from an internal buffer, which in turn
(when empty) draws data from the actual file.

This is convoluted, I know, but it's the way the cfile library is
intended to work.  We don't simply give the user a chunk of data and
say "you sort it out", letting them drive when they need more file
to be decompressed - we provide a regular stdio-like interface.  That
means the caller doesn't need to know about this buffering and
decompression, it happens transparently underneath them.