This to-do is now done.
I misunderstood Tridge - he wasn't talking about trying to interface with
stdio.h, he was talking about replacing the if/then cases in each function
with a functional table of the type that Peter Miller then explained to
me.  As usual, Tridge is way ahead of everyone else :-)  But this is now
implemented and working.